% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drsurv_crt.R
\name{summary.dr_surv}
\alias{summary.dr_surv}
\title{Summary for Doubly Robust Survival Estimates}
\usage{
\method{summary}{dr_surv}(object, alpha = 0.05)
}
\arguments{
\item{object}{A list returned from `drsurv_crt()`, containing survival estimates.}

\item{alpha}{A numeric value (default `0.05`) specifying the significance level
for confidence intervals (CI). The function calculates `(1 - alpha)%` CI.}
}
\value{
A list with summary statistics:
  - `$summary_s_c`: Summary of **cluster-level survival estimates** with confidence intervals.
  - `$summary_s_i`: Summary of **individual-level survival estimates** with confidence intervals.
}
\description{
This function provides a summary of survival estimates from `drsurv_crt()`,
including confidence intervals based on the specified significance level.
}
